+++ packages/metadata-tools/src/config-metadata.ts
@@ -1,8 +1,19 @@
 import type { OmniEdgeHardhat } from '@layerzerolabs/devtools-evm-hardhat'
 import type { OAppEdgeConfig } from '@layerzerolabs/ua-devtools'
-import { IMetadata } from './types'
+import { BlockConfirmationsDefinition, BlockConfirmationsType, IMetadata } from './types'
 import { TwoWayConfig } from './types'
-import { METADATA_URL } from './constants'
+import {
+    METADATA_KEY_EVM_BLOCKED_MESSAGE,
+    METADATA_KEY_SOLANA_BLOCKED_MESSAGE,
+    METADATA_KEY_RECEIVE_LIBRARY,
+    METADATA_KEY_SEND_LIBRARY,
+    METADATA_URL,
+    MSG_LIB_BLOCK_RECEIVE_ONLY,
+    MSG_LIB_BLOCK_SEND_AND_RECEIVE,
+    MSG_LIB_BLOCK_SEND_ONLY,
+    NIL_DVN_COUNT,
+} from './constants'
+import { getAddress } from '@ethersproject/address'
 
 function getEndpointIdDeployment(eid: number, metadata: IMetadata) {
     const srcEidString = eid.toString()
@@ -127,17 +138,49 @@ function resolveExecutorForDeployment(
     return defaultExecutor
 }
 
+// Helper to resolve library metadata key based on operation and chain
+function resolveLibraryMetadataKey(isSend: boolean, isBlocked: boolean, deployment: { chainKey: string }) {
+    if (isBlocked) {
+        if (isSolanaDeployment(deployment)) {
+            return METADATA_KEY_SOLANA_BLOCKED_MESSAGE
+        }
+
+        return METADATA_KEY_EVM_BLOCKED_MESSAGE
+    }
+    return isSend ? METADATA_KEY_SEND_LIBRARY : METADATA_KEY_RECEIVE_LIBRARY
+}
+
+// Helper to assert presence of a metadata key on a deployment
+function assertHasKey(deployment: Record<string, any>, key: string, eid: number) {
+    if (!deployment[key]) {
+        throw new Error(`Can't find ${key} for endpoint with eid: "${eid}".`)
+    }
+}
+
+function isBlocked(blockConfirmationsDefinition: BlockConfirmationsDefinition | undefined, isSend: boolean): boolean {
+    const keys = [MSG_LIB_BLOCK_SEND_AND_RECEIVE, isSend ? MSG_LIB_BLOCK_SEND_ONLY : MSG_LIB_BLOCK_RECEIVE_ONLY]
+
+    return Boolean(
+        blockConfirmationsDefinition &&
+            typeof blockConfirmationsDefinition === 'object' &&
+            keys.includes(blockConfirmationsDefinition[1])
+    )
+}
+
+const getLibraryAddress = (deployment: any, metadataKey: string) =>
+    isSolanaDeployment(deployment) ? deployment[metadataKey].address : getAddress(deployment[metadataKey].address)
+
 export async function translatePathwayToConfig(
     pathway: TwoWayConfig,
     metadata: IMetadata
 ): Promise<OmniEdgeHardhat<OAppEdgeConfig | undefined>[]> {
     const configs: OmniEdgeHardhat<OAppEdgeConfig | undefined>[] = []
 
-    const sourceContract = pathway[0]
-    const destinationContract = pathway[1]
+    const AContract = pathway[0]
+    const BContract = pathway[1]
     const [requiredDVNs, optionalDVNConfig] = pathway[2]
-    const [sourceToDestinationConfirmations, destinationToSourceConfirmations] = pathway[3]
-    const [enforcedOptionsSrcToDst, enforcedOptionsDstToSrc] = pathway[4]
+    const [AToBConfirmationsDefinition, BToAConfirmationsDefinition] = pathway[3]
+    const [enforcedOptionsAToB, enforcedOptionsBToA] = pathway[4]
     const customExecutor = pathway[5]
 
     const optionalDVNs = optionalDVNConfig[0]
@@ -147,123 +190,131 @@ export async function translatePathwayToConfig(
         throw new Error(`Optional DVN threshold is greater than the number of optional DVNs.`)
     }
 
-    const sourceLZDeployment = getEndpointIdDeployment(sourceContract.eid, metadata)
-    const destinationLZDeployment = getEndpointIdDeployment(destinationContract.eid, metadata)
+    const ALZDeployment = getEndpointIdDeployment(AContract.eid, metadata)
+    const BLZDeployment = getEndpointIdDeployment(BContract.eid, metadata)
 
-    const sourceExecutor = resolveExecutorForDeployment(
-        customExecutor,
-        sourceLZDeployment,
-        metadata,
-        sourceContract.eid
-    )
+    const AExecutor = resolveExecutorForDeployment(customExecutor, ALZDeployment, metadata, AContract.eid)
 
-    const sourceRequiredDVNs = DVNsToAddresses(requiredDVNs, sourceLZDeployment.chainKey, metadata)
-    const destinationRequiredDVNs = DVNsToAddresses(requiredDVNs, destinationLZDeployment.chainKey, metadata)
+    const ARequiredDVNs = DVNsToAddresses(requiredDVNs, ALZDeployment.chainKey, metadata)
+    const BRequiredDVNs = DVNsToAddresses(requiredDVNs, BLZDeployment.chainKey, metadata)
+    const requiredDVNCount = requiredDVNs.length > 0 ? requiredDVNs.length : NIL_DVN_COUNT
 
-    let sourceOptionalDVNs: string[] = []
-    let destinationOptionalDVNs: string[] = []
+    let AOptionalDVNs: string[] = []
+    let BOptionalDVNs: string[] = []
 
     if (optionalDVNs) {
-        sourceOptionalDVNs = DVNsToAddresses(optionalDVNs, sourceLZDeployment.chainKey, metadata)
-        destinationOptionalDVNs = DVNsToAddresses(optionalDVNs, destinationLZDeployment.chainKey, metadata)
+        AOptionalDVNs = DVNsToAddresses(optionalDVNs, ALZDeployment.chainKey, metadata)
+        BOptionalDVNs = DVNsToAddresses(optionalDVNs, BLZDeployment.chainKey, metadata)
     }
 
-    if (!sourceLZDeployment.sendUln302 || !sourceLZDeployment.receiveUln302 || !sourceLZDeployment.executor) {
-        throw new Error(
-            `Can't find sendUln302, receiveUln302 or executor for source endpoint with eid: "${sourceContract.eid}".`
-        )
-    }
+    const AToBConfirmations: BlockConfirmationsType = ['bigint', 'number'].includes(typeof AToBConfirmationsDefinition)
+        ? AToBConfirmationsDefinition
+        : AToBConfirmationsDefinition[0]
+    const BToAConfirmations: BlockConfirmationsType | undefined = ['bigint', 'number'].includes(
+        typeof BToAConfirmationsDefinition
+    )
+        ? BToAConfirmationsDefinition
+        : BToAConfirmationsDefinition?.[0]
 
-    if (
-        !destinationLZDeployment.sendUln302 ||
-        !destinationLZDeployment.receiveUln302 ||
-        !destinationLZDeployment.executor
-    ) {
-        throw new Error(
-            `Can't find sendUln302, receiveUln302 or executor for destination endpoint with eid: "${destinationContract.eid}".`
-        )
-    }
+    const blockSendAToB = isBlocked(AToBConfirmationsDefinition, true)
+    const blockReceiveAToB = isBlocked(AToBConfirmationsDefinition, false)
+    const blockSendBToA = isBlocked(BToAConfirmationsDefinition, true)
+    const blockReceiveBToA = isBlocked(BToAConfirmationsDefinition, false)
 
-    const sourceToDestinationConfig: OmniEdgeHardhat<OAppEdgeConfig> = {
-        from: sourceContract,
-        to: destinationContract,
+    const sendLibraryAToBMetadataKey = resolveLibraryMetadataKey(true, blockSendAToB, ALZDeployment)
+    const receiveLibraryAToBMetadataKey = resolveLibraryMetadataKey(false, blockReceiveAToB, BLZDeployment)
+    const sendLibraryBToAMetadataKey = resolveLibraryMetadataKey(true, blockSendBToA, BLZDeployment)
+    const receiveLibraryBToAMetadataKey = resolveLibraryMetadataKey(false, blockReceiveBToA, ALZDeployment)
+
+    assertHasKey(ALZDeployment, sendLibraryAToBMetadataKey, AContract.eid)
+    assertHasKey(BLZDeployment, receiveLibraryAToBMetadataKey, BContract.eid)
+    assertHasKey(BLZDeployment, sendLibraryBToAMetadataKey, BContract.eid)
+    assertHasKey(ALZDeployment, receiveLibraryBToAMetadataKey, AContract.eid)
+
+    const sendLibraryAToB = getLibraryAddress(ALZDeployment, sendLibraryAToBMetadataKey)
+    const receiveLibraryAToB = getLibraryAddress(BLZDeployment, receiveLibraryAToBMetadataKey)
+    const sendLibraryBToA = getLibraryAddress(BLZDeployment, sendLibraryBToAMetadataKey)
+    const receiveLibraryBToA = getLibraryAddress(ALZDeployment, receiveLibraryBToAMetadataKey)
+
+    const AToBConfig: OmniEdgeHardhat<OAppEdgeConfig> = {
+        from: AContract,
+        to: BContract,
         config: {
-            sendLibrary: sourceLZDeployment.sendUln302.address,
+            sendLibrary: sendLibraryAToB,
             receiveLibraryConfig: {
-                receiveLibrary: sourceLZDeployment.receiveUln302.address,
+                receiveLibrary: receiveLibraryBToA,
                 gracePeriod: BigInt(0),
             },
             sendConfig: {
                 executorConfig: {
                     maxMessageSize: 10000,
-                    executor: sourceExecutor,
+                    executor: AExecutor,
                 },
                 ulnConfig: {
-                    confirmations: BigInt(sourceToDestinationConfirmations),
-                    requiredDVNs: sourceRequiredDVNs,
-                    optionalDVNs: sourceOptionalDVNs,
+                    confirmations: BigInt(AToBConfirmations),
+                    requiredDVNs: ARequiredDVNs,
+                    requiredDVNCount,
+                    optionalDVNs: AOptionalDVNs,
                     optionalDVNThreshold,
                 },
             },
-            enforcedOptions: enforcedOptionsSrcToDst,
+            enforcedOptions: enforcedOptionsAToB,
         },
     }
 
-    const destinationToSourceConfig: OmniEdgeHardhat<OAppEdgeConfig> = {
-        from: destinationContract,
-        to: sourceContract,
+    const BToAConfig: OmniEdgeHardhat<OAppEdgeConfig> = {
+        from: BContract,
+        to: AContract,
         config: {
-            sendLibrary: destinationLZDeployment.sendUln302.address,
+            sendLibrary: sendLibraryBToA,
             receiveLibraryConfig: {
-                receiveLibrary: destinationLZDeployment.receiveUln302.address,
+                receiveLibrary: receiveLibraryAToB,
                 gracePeriod: BigInt(0),
             },
             receiveConfig: {
                 ulnConfig: {
-                    confirmations: BigInt(sourceToDestinationConfirmations),
-                    requiredDVNs: destinationRequiredDVNs,
-                    optionalDVNs: destinationOptionalDVNs,
+                    confirmations: BigInt(AToBConfirmations),
+                    requiredDVNs: BRequiredDVNs,
+                    requiredDVNCount,
+                    optionalDVNs: BOptionalDVNs,
                     optionalDVNThreshold,
                 },
             },
         },
     }
 
-    if (destinationToSourceConfirmations) {
-        const destinationExecutor = resolveExecutorForDeployment(
-            customExecutor,
-            destinationLZDeployment,
-            metadata,
-            destinationContract.eid
-        )
+    if (BToAConfirmations) {
+        const BExecutor = resolveExecutorForDeployment(customExecutor, BLZDeployment, metadata, BContract.eid)
 
-        sourceToDestinationConfig.config.receiveConfig = {
+        AToBConfig.config.receiveConfig = {
             ulnConfig: {
-                confirmations: BigInt(destinationToSourceConfirmations),
-                requiredDVNs: sourceRequiredDVNs,
-                optionalDVNs: sourceOptionalDVNs,
+                confirmations: BigInt(BToAConfirmations),
+                requiredDVNs: ARequiredDVNs,
+                requiredDVNCount,
+                optionalDVNs: AOptionalDVNs,
                 optionalDVNThreshold,
             },
         }
 
-        destinationToSourceConfig.config.enforcedOptions = enforcedOptionsDstToSrc
+        BToAConfig.config.enforcedOptions = enforcedOptionsBToA
 
-        destinationToSourceConfig.config.sendConfig = {
+        BToAConfig.config.sendConfig = {
             executorConfig: {
                 maxMessageSize: 10000,
-                executor: destinationExecutor,
+                executor: BExecutor,
             },
             ulnConfig: {
-                confirmations: BigInt(destinationToSourceConfirmations),
-                requiredDVNs: destinationRequiredDVNs,
-                optionalDVNs: destinationOptionalDVNs,
+                confirmations: BigInt(BToAConfirmations),
+                requiredDVNs: BRequiredDVNs,
+                requiredDVNCount,
+                optionalDVNs: BOptionalDVNs,
                 optionalDVNThreshold,
             },
         }
     }
 
-    configs.push(sourceToDestinationConfig)
-    configs.push(destinationToSourceConfig)
+    configs.push(AToBConfig)
+    configs.push(BToAConfig)
 
     return configs
 }