+++ packages/ua-devtools/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @layerzerolabs/ua-devtools
 
+## 5.0.1
+
+### Patch Changes
+
+- db3a1d4: add block msg lib support on evm and optional dvns support on evm and solana
+- db3a1d4: Updates simple config to support blocked messageLib on Solana and EVM
+- Updated dependencies [db3a1d4]
+- Updated dependencies [db3a1d4]
+  - @layerzerolabs/protocol-devtools@3.0.1
+
 ## 5.0.0
 
 ### Patch Changes